LayoutTests/http/tests/inspector/network/x-frame-options-expected.txt

CONSOLE MESSAGE: Refused to display 'http://127.0.0.1:8000/inspector/network/resources/x-frame-options.php?option=DENY' in a frame because it set 'X-Frame-Options' to 'DENY'.
Tests for various X-Frame-Options headers.


== Running test suite: Network.XFrameOptions
-- Running test case: Network.XFrameOptions.Denied
PASS: X-Frame-Options headers should match.

-- Running test case: Network.XFrameOptions.Sameorigin
PASS: X-Frame-Options headers should match.
